+++ .gitignore
@@ -5,22 +5,13 @@ __pycache__/
 .ipynb_checkpoints
 */.ipynb_checkpoints/*
 
-MVAD/*
 
 .vscode/*
 */.vscode/*
-!.vscode/settings.json
-!.vscode/tasks.json
-!.vscode/launch.json
-!.vscode/extensions.json
 .history
 
-lib/Untitled*.ipynb
-
-lib/__pycache__/*
-
+data/*
 logs/*
 checkpoint/*
 
-
-lib/.fuse_hidden*
+lib/*
